+++ packages/elements/table-ui/src/index.ts
@@ -3,7 +3,9 @@
  */
 
 export * from './table.atoms';
-export * from './TableElement/index';
 export * from './TableCellElement/index';
+export * from './TableElement/index';
+export * from './TablePopover/index';
+export * from './TableRowElement/index';
 export * from './ToolbarTable/index';
 export * from './hooks/index';