+++ templates/menu/src/index.tsx
@@ -1,0 +1,16 @@
+// export {
+//   applyHOC,
+//   applyStatics,
+//   useMenuState,
+//   ControlledMenu,
+//   SubMenu,
+//   MenuItem,
+//   FocusableItem,
+//   MenuDivider,
+//   MenuHeader,
+//   MenuGroup,
+//   MenuRadioGroup,
+// } from '@szhsin/react-menu';
+
+export * from './Menu';
+export * from './MenuButton';