+++ packages/elements/basic-elements/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-basic-elements
 
+## 6.2.0
+
 ## 6.1.0
 
 ## 6.0.0