+++ packages/elements/table/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-table
 
+## 6.2.0
+
 ## 6.1.0
 
 ### Minor Changes