+++ packages/marks/font-ui/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-font-ui
 
+## 6.2.0
+
 ## 6.1.0
 
 ## 6.0.0