+++ packages/serializers/ast-serializer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-ast-serializer
 
+## 6.2.0
+
 ## 6.1.0
 
 ## 6.0.0