+++ packages/ui/combobox/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-combobox
 
+## 6.2.0
+
 ## 6.1.0
 
 ### Patch Changes